Tech demo of first game

+++ asthma/src/com/groupc/game/GameObject.java
@@ -23,7 +23,7 @@ public class GameObject
 	 * @param width - width of object
 	 * @param height - width of object
 	 */
-	public GameObject(float x, float y, int width, int height)
+	public GameObject(float x, float y, float width, float height)
 	{
 		position = new Vector(x, y);
 		bounds = new Rectangle(x - width/2, y - height/2, width, height);
